requirements.txt
playwright
beautifulsoup4
pytesseract
Pillow
openai
python-dotenv
streamlit
